Fixed issue where bindings weren't checked before unifying
Now it will check that bindings being merged that have the same
variable name also have the same value or else it will fail.  This
was causing the "append" test to misbehave.

An issue has also been fixed where builtin predicate evaluation was
causing an infinite loop when backtracking, and evaluting the builtin
multiple times.